Completed the refactor of the codebase
