+++ src/dto/user_info.rs
@@ -1,4 +1,4 @@
-use crate::auth::{Role, UserInfo};
+use crate::models::user_info::{Role, UserInfo};
 use crate::models::datetime::Datetime;
 use crate::utils::create_dto::create_dto;
