+++ src/endpoints/api/user/update.rs
@@ -4,7 +4,7 @@ use crate::extractors::AuthenticatedToken;
 use crate::extractors::IntoSession;
 use crate::generate_endpoint;
 use crate::services::user::update::*;
-use crate::services::user::utils::get_user_by_token;
+use crate::services::user::get::get_user_by_token;
 use crate::state::AppState;
 use actix_web::{web, HttpResponse};
 use utoipa::ToSchema;
