+++ src/services/auth_for/mod.rs
@@ -1,0 +1,3 @@
+pub mod create;
+
+pub(crate) use create::*;
