+++ src/services/mod.rs
@@ -4,3 +4,4 @@
 pub(crate) mod health;
 pub(crate) mod token;
 pub(crate) mod user;
+pub(crate) mod auth_for;
